feat(ui): BundleModules - show all entries
