--- packages/ui/src/components/bundle-chunk-modules/bundle-chunk-modules.module.css
@@ -1,53 +1,0 @@
-.title {
-  color: var(--color-text-light);
-}
-
-.header {
-  padding: 0 0 var(--space-small);
-}
-
-.headerTitle {
-  margin: 0;
-  font-size: 1em;
-  overflow-wrap: break-word;
-}
-
-.headerTitleName {
-  display: block;
-  font-size: var(--size-large);
-}
-
-.headerTitleId {
-  display: block;
-  color: var(--color-text-light);
-  font-size: var(--size-small);
-  line-height: var(--space-small);
-}
-
-.toolbar {
-  border-bottom: 1px solid var(--color-outline);
-}
-
-.toolbarSearch {
-  width: 100%;
-  max-width: 360px;
-}
-
-.nameTooltip {
-  text-align: left;
-}
-
-.nameTooltipTitle {
-  margin: 0;
-}
-
-.nameTooltipText {
-  display: block;
-  width: 100%;
-  max-width: 320px;
-  min-width: 280px;
-}
-
-.nameTooltipItem + .nameTooltipItem {
-  margin-top: var(--space-small);
-}
