--- packages/ui/src/components/bundle-chunk-modules/index.js
@@ -1,4 +1,0 @@
-import enhance from './bundle-chunk-modules.container';
-import { BundleChunkModules as Component } from './bundle-chunk-modules';
-
-export const BundleChunkModules = enhance(Component);
