+++ packages/ui/src/components/bundle-modules/bundle-modules.stories.jsx
@@ -1,38 +1,71 @@
 import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { createJobs } from '@bundle-stats/utils';
-import { merge, set } from 'lodash';
 
 import baselineStats from '../../../__mocks__/webpack-stats.baseline.json';
 import currentStats from '../../../__mocks__/webpack-stats.current.json';
 import { getWrapperDecorator } from '../../stories';
-import { BundleModules } from './bundle-modules';
+import { BundleModules } from '.';
 
-const [currentJob, baselineJob] = createJobs([
+const JOBS = createJobs([
   { webpack: currentStats },
   { webpack: baselineStats },
 ]);
+
+// eslint-disable-next-line no-unused-variables
+const [CURRENT_JOB, BASELINE_JOB]= JOBS;
 
 const stories = storiesOf('Components/BundleModules', module);
 stories.addDecorator(getWrapperDecorator());
 
 stories.add('default', () => (
-  <BundleModules jobs={[baselineJob]} />
+  <BundleModules jobs={[BASELINE_JOB]} />
 ));
 
-stories.add('multiple runs', () => (
-  <BundleModules jobs={[currentJob, baselineJob]} />
+stories.add('multiple jobs', () => (
+  <BundleModules jobs={JOBS} />
 ));
+
+// stories.add('empty modules', () => (
+//   <BundleChunkModules jobs={JOBS.map((job) => merge({}, job, { metrics: { webpack: { modules: {} } } }))} />
+// ));
+//
+// stories.add('empty filtered modules', () => (
+//   <BundleChunkModules
+//     runs={JOBS}
+//     items={[
+//       {
+//         key: 'module-a',
+//         label: 'module-a',
+//         biggerIsBetter: false,
+//         changed: false,
+//         runs: [
+//           {
+//             name: 'module-a',
+//             value: 25,
+//             displayValue: '25B',
+//             delta: 0,
+//             deltaPercentage: 0,
+//             displayDelta: '0B',
+//             displayDeltaPercentage: '0%',
+//           },
+//           {
+//             name: 'module-a',
+//             value: 25,
+//             displayValue: '25B',
+//             delta: 0,
+//             deltaPercentage: 0,
+//             displayDelta: '0B',
+//             displayDeltaPercentage: '0%',
+//           },
+//         ],
+//       },
+//     ]}
+//   />
+// ));
 
 const JOBS_EMPTY_BASELINE = createJobs([{ webpack: currentStats }, null]);
 
-stories.add('empty baseline', () => <BundleModules jobs={JOBS_EMPTY_BASELINE} />);
-
-stories.add('no modules', () => (
-  <BundleModules
-    jobs={[
-      set(merge({}, currentJob), 'metrics.webpack.modules', {}),
-      set(merge({}, baselineJob), 'metrics.webpack.modules', {}),
-    ]}
-  />
+stories.add('empty baseline', () => (
+  <BundleModules jobs={JOBS_EMPTY_BASELINE} />
 ));
