+++ packages/ui/src/components/bundle-modules/index.js
@@ -1,1 +1,4 @@
-export { BundleModules } from './bundle-modules';
+import enhance from './bundle-modules.container';
+import { BundleModules as Component } from './bundle-modules';
+
+export const BundleModules = enhance(Component);
