+++ packages/ui/src/components/index.js
@@ -2,7 +2,6 @@ export { BrowsertimeTable } from './browsertime-table';
 export { BundleAssets } from './bundle-assets';
 export { BundleAssetsTotalsTable } from './bundle-assets-totals-table';
 export { BundleAssetsTotalsChartBars } from './bundle-assets-totals-chart-bars';
-export { BundleChunkModules } from './bundle-chunk-modules';
 export { BundlePackages } from './bundle-packages';
 export { BundleModules } from './bundle-modules';
 export { ComponentLink } from './component-link';
